Code/minsight_runner_ros/requirements.txt
cv_bridge==3.2.1
matplotlib==3.5.1
numpy==1.21.5
Pillow
rclpy==3.3.8
rosbags==0.9.15
torch==2.0.1
torchvision==0.15.2
